refactor(api): Consolidate endpoints
- Put sessions & accounts modules together. Remove session API module.
- Remove outdated comments
